Create swap.py
Modified the above program so that what’s in the variable word1 is now in the variable word2 and what’s in variable word2 is now in variable word1. I was not allowed to remove any lines of code from this program. I could only add assignment statements between lines 3 and 6 to get the desired result.